+++ lib/job-iteration/csv_enumerator.rb
@@ -49,7 +49,14 @@ module JobIteration
     private
 
     def count_of_rows_in_file
-      filepath = @csv.path
+      # TODO: Remove rescue for NoMethodError when Ruby 2.6 is no longer supported.
+      begin
+        filepath = @csv.path
+      rescue NoMethodError
+        return
+      end
+
+      # Behaviour of CSV#path changed in Ruby 2.6.3 (returns nil instead of raising NoMethodError)
       return unless filepath
 
       count = %x(wc -l < #{filepath}).strip.to_i